--- src/pages/Dashboard.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Dashboard() {
-  return (
-    <div>Dashboard</div>
-  )
-}
