+++ src/user-auth/LogIn.js
@@ -1,31 +1,26 @@
 import React, { useRef, useState } from 'react'
 import { Alert, Form, Button, Card, Container } from 'react-bootstrap'
-import { useAuth } from '../components/user-auth/contexts/AuthContexts'
+import { useAuth } from './contexts/AuthContexts'
 import { Link, useNavigate } from 'react-router-dom'
 
-export default function SignUp() {
+export default function LogIn() {
     const emailRef = useRef()
     const passwordRef = useRef()
-    const passwordConfirmationRef = useRef()
-    const { signup } = useAuth()
-    const { error, setError } = useState('')
-    const { loading, setLoading } = useState(false)
+    const { login } = useAuth()
+    const [ error, setError ] = useState('')
+    const [ loading, setLoading ] = useState(false)
     const navigate = useNavigate()
 
     async function handleSubmit(e) {
         e.preventDefault()
 
-        if (passwordRef.current.value !== passwordConfirmationRef.current.value) {
-            return setError('Passwords do not match')
-        }
-
         try {
             setError('')
             setLoading(true)
-            await signup(emailRef.current.value, passwordRef.current.value)
-            navigate('/')
+            await login(emailRef.current.value, passwordRef.current.value)
+            navigate('/dash')
         } catch {
-            setError('Failed to create an account')
+            setError('Failed to log in')
         }
         setLoading(false)
     }
@@ -35,7 +30,7 @@ export default function SignUp() {
             <Container className='mt-5' style={{ maxWidth: "600px" }}>
                 <Card>
                     <Card.Body>
-                        <h2 className='text-center mb-4'>Sign Up</h2>
+                        <h2 className='text-center mb-4'>Log In</h2>
                         {error && <Alert variant="danger">{error}</Alert>}
                         <Form onSubmit={handleSubmit}>
                             <Form.Group id='email'>
@@ -46,18 +41,14 @@ export default function SignUp() {
                                 <Form.Label>Password</Form.Label>
                                 <Form.Control type='password' ref={passwordRef} required></Form.Control>
                             </Form.Group>
-                            <Form.Group id='password-confirm'>
-                                <Form.Label>Confirm Password</Form.Label>
-                                <Form.Control type='password' ref={passwordConfirmationRef} required></Form.Control>
-                            </Form.Group>
                             <Button disabled={loading} className='w-100 mt-2' type='submit'>
-                                Sign Up
+                                Log In
                             </Button>
                         </Form>
                     </Card.Body>
                 </Card>
                 <div className='w-100 text-center mt-2'>
-                    Already have an account? <Link to='/login'>Log In</Link>
+                    Need an account? <Link to='/signup'>Sign Up</Link>
                 </div>
             </Container>
         </>
