+++ src/user-auth/SignUp.js
@@ -1,26 +1,31 @@
 import React, { useRef, useState } from 'react'
 import { Alert, Form, Button, Card, Container } from 'react-bootstrap'
-import { useAuth } from '../components/user-auth/contexts/AuthContexts'
+import { useAuth } from './contexts/AuthContexts'
 import { Link, useNavigate } from 'react-router-dom'
 
-export default function LogIn() {
+export default function SignUp() {
     const emailRef = useRef()
     const passwordRef = useRef()
-    const { login } = useAuth()
-    const { error, setError } = useState('')
-    const { loading, setLoading } = useState(false)
+    const passwordConfirmationRef = useRef()
+    const { signup } = useAuth()
+    const [ error, setError ] = useState('')
+    const [ loading, setLoading ] = useState(false)
     const navigate = useNavigate()
 
     async function handleSubmit(e) {
         e.preventDefault()
 
+        if (passwordRef.current.value !== passwordConfirmationRef.current.value) {
+            return setError('Passwords do not match')
+        }
+
         try {
             setError('')
             setLoading(true)
-            await login(emailRef.current.value, passwordRef.current.value)
-            navigate('/')
+            await signup(emailRef.current.value, passwordRef.current.value)
+            navigate('/dash')
         } catch {
-            setError('Failed to log in')
+            setError('Failed to create an account')
         }
         setLoading(false)
     }
@@ -30,7 +35,7 @@ export default function LogIn() {
             <Container className='mt-5' style={{ maxWidth: "600px" }}>
                 <Card>
                     <Card.Body>
-                        <h2 className='text-center mb-4'>Log In</h2>
+                        <h2 className='text-center mb-4'>Sign Up</h2>
                         {error && <Alert variant="danger">{error}</Alert>}
                         <Form onSubmit={handleSubmit}>
                             <Form.Group id='email'>
@@ -41,14 +46,18 @@ export default function LogIn() {
                                 <Form.Label>Password</Form.Label>
                                 <Form.Control type='password' ref={passwordRef} required></Form.Control>
                             </Form.Group>
+                            <Form.Group id='password-confirm'>
+                                <Form.Label>Confirm Password</Form.Label>
+                                <Form.Control type='password' ref={passwordConfirmationRef} required></Form.Control>
+                            </Form.Group>
                             <Button disabled={loading} className='w-100 mt-2' type='submit'>
-                                Log In
+                                Sign Up
                             </Button>
                         </Form>
                     </Card.Body>
                 </Card>
                 <div className='w-100 text-center mt-2'>
-                    Need an account? <Link to='/signup'>Sign Up</Link>
+                    Already have an account? <Link to='/login'>Log In</Link>
                 </div>
             </Container>
         </>
